+++ lib/atlas.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Atlas
+  class << self
+    def configuration
+      @configuration ||= Configuration.new
+    end
+
+    def configure
+      yield(configuration) if block_given?
+    end
+
+    def reset_configuration!
+      @configuration = nil
+    end
+  end
+end